requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.2.0
autograd==1.6.2
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164373/work
cachetools==5.2.0
captum==0.5.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1625835307225/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
cloudpickle==2.2.0
colorama==0.4.6
colored==1.4.4
comm==0.2.1
CORnet @ file:///oak/stanford/groups/menon/projects/astrock/2021_common/scripts/packages/CORnet
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967134882/work
cvxopt==1.3.2
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython==0.29.35
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
dm-tree==0.1.7
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
ffmpeg==1.4
fire==0.4.0
flatbuffers==2.0.7
fonttools==4.25.0
fqdn==1.5.1
freetype-py==2.3.0
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.2
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.1
h11==0.14.0
h5py==3.7.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
hyperopt==0.2.7
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.29.0
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.5.12
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1653292039266/work
libclang==14.0.6
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
littleutils==0.2.2
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib @ file:///opt/conda/conda-bld/matplotlib-suite_1660167928326/work
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==2.8.8
nibabel==4.0.2
nilearn==0.10.2
notebook==7.0.7
notebook_shim==0.2.3
numba==0.57.1
numexpr @ file:///opt/conda/conda-bld/numexpr_1656940300424/work
numpy==1.22.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.7.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.4
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1621287979692/work
platformdirs==4.2.0
pluggy==1.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyDeprecate==0.3.1
Pygments==2.17.2
pymanopt==2.2.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytest==7.1.3
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
python-slugify==7.0.0
pytorch-lightning==1.8.0.post1
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyYAML==6.0
pyzmq==24.0.1
receptive-field @ file:///oak/stanford/groups/menon/projects/astrock/test/receptive_field
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
reservoirpy @ file:///oak/stanford/groups/menon/projects/astrock/2021_common/scripts/packages/reservoirpy
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.1.3
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.2
sip==4.19.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
sorcery==0.2.2
soupsieve==2.5
stack-data==0.6.3
statannotations==0.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
terminado==0.18.0
text-unidecode==1.3
tf-slim==1.1.0
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.10.0
torchvision==0.13.1
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets==5.14.1
typeguard==2.13.3
types-python-dateutil==2.8.19.20240106
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.9.0
